Re enable tests
